VLM/zero_shot_classification/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
open-clip-torch>=2.20.0
Pillow>=9.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pandas>=2.0.0
